requirements.txt
typesense
pandas
streamlit
python-dotenv
requests